Update to v1 of reflection schema
The gRPC folks graduated server reflection to V1 without making any
changes to the schema. This PR updates our schema to v1, fixes the
protobuf package path to match the directory, and retains the gRPC
license (as we should).

Many tools still use the v1alpha1 API, so this PR also adds a second
exported handler constructor for the alpha API. I considered doing this
with HTTP middleware, but it seems more straighforward to just offer two
handlers.