Add grpcreflect.Client (#42)
This provides an RPC client interface for the gRPC
Server Reflection service. It does not expose any of
the generated base types for the service, and instead
offers a higher-level interface for interacting with
the stream. The client automatically supports both
v1 and v1alpha of the service, trying v1 first and then
using v1alpha if that results in an "unimplemented"
error.