Don't send placeholder files; improve caching of sent files (#66)
Placeholder files represent missing dependencies in the
protoreflect library. When the reflection service was asked
serialize a file and its dependencies, if any of its imports
were missing from the global registry, they would appear
in the form of a placeholder file, which would then be
serialized in the response. However, the file descriptor
that was serialized would be invalid and unusable by the
client. This fixes that issue: placeholder files are not sent
to the client and instead interpreted as missing files.

While writing new tests for this, I also discovered that we
were caching sent files incorrectly. The code was using
`FileDescriptor.FullName`, but this returns the file's package,
not its name or path. .This is not enough to uniquely identify
a file, so the code was sending back only one dependency
per package instead of the entire graph, which then required
the client ask for the others, one-at-a-time. This commit
fixes that as well.